Tests/LibWeb/Text/expected/URL/URL-operations-properties.txt

writable: true
configurable: true
enumerable: true
value (the function itself): function canParse() { [native code] }
